requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 24.9.2
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
accelerate=1.0.1=pypi_0
aiohappyeyeballs=2.4.3=pypi_0
aiohttp=3.10.10=pypi_0
aiosignal=1.3.1=pypi_0
antlr4-python3-runtime=4.9.3=pypi_0
async-timeout=4.0.3=pypi_0
attrs=24.2.0=pypi_0
audioread=3.0.1=pypi_0
av=13.1.0=pypi_0
beautifulsoup4=4.12.3=pypi_0
bzip2=1.0.8=h5eee18b_6
ca-certificates=2024.9.24=h06a4308_0
certifi=2024.8.30=pypi_0
cffi=1.17.1=pypi_0
charset-normalizer=3.4.0=pypi_0
clip=1.0=pypi_0
contourpy=1.3.0=pypi_0
cycler=0.12.1=pypi_0
datasets=3.0.2=pypi_0
decorator=5.1.1=pypi_0
decord=0.6.0=pypi_0
diffusers=0.31.0=pypi_0
dill=0.3.8=pypi_0
efficientnet-pytorch=0.7.1=pypi_0
einops=0.8.0=pypi_0
filelock=3.16.1=pypi_0
fonttools=4.54.1=pypi_0
frozenlist=1.4.1=pypi_0
fsspec=2024.9.0=pypi_0
ftfy=6.3.1=pypi_0
fvcore=0.1.5.post20221221=pypi_0
gdown=5.2.0=pypi_0
huggingface-hub=0.25.0=pypi_0
idna=3.10=pypi_0
imageio=2.36.0=pypi_0
imageio-ffmpeg=0.5.1=pypi_0
importlib-metadata=8.5.0=pypi_0
iopath=0.1.10=pypi_0
jaxtyping=0.2.34=pypi_0
jinja2=3.1.4=pypi_0
joblib=1.4.2=pypi_0
kiwisolver=1.4.7=pypi_0
kornia=0.7.4=pypi_0
kornia-rs=0.1.7=pypi_0
lazy-loader=0.4=pypi_0
ld_impl_linux-64=2.40=h12ee557_0
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
librosa=0.10.2.post1=pypi_0
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
lightning-utilities=0.11.9=pypi_0
llvmlite=0.43.0=pypi_0
markupsafe=3.0.2=pypi_0
matplotlib=3.9.2=pypi_0
moviepy=2.1.2=pypi_0
mpmath=1.3.0=pypi_0
msgpack=1.1.0=pypi_0
multidict=6.1.0=pypi_0
multiprocess=0.70.16=pypi_0
ncurses=6.4=h6a678d5_0
networkx=3.4.2=pypi_0
numba=0.60.0=pypi_0
numpy=1.24.1=pypi_0
nvidia-cublas-cu11=11.10.3.66=pypi_0
nvidia-cublas-cu12=12.1.3.1=pypi_0
nvidia-cuda-cupti-cu12=12.1.105=pypi_0
nvidia-cuda-nvrtc-cu11=11.7.99=pypi_0
nvidia-cuda-nvrtc-cu12=12.1.105=pypi_0
nvidia-cuda-runtime-cu11=11.7.99=pypi_0
nvidia-cuda-runtime-cu12=12.1.105=pypi_0
nvidia-cudnn-cu11=8.5.0.96=pypi_0
nvidia-cudnn-cu12=8.9.2.26=pypi_0
nvidia-cufft-cu12=11.0.2.54=pypi_0
nvidia-curand-cu12=10.3.2.106=pypi_0
nvidia-cusolver-cu12=11.4.5.107=pypi_0
nvidia-cusparse-cu12=12.1.0.106=pypi_0
nvidia-nccl-cu12=2.19.3=pypi_0
nvidia-nvjitlink-cu12=12.4.127=pypi_0
nvidia-nvtx-cu12=12.1.105=pypi_0
omegaconf=2.3.0=pypi_0
opencv-python=4.10.0.84=pypi_0
openssl=3.0.15=h5eee18b_0
packaging=24.1=pypi_0
pandas=2.2.3=pypi_0
parameterized=0.9.0=pypi_0
pillow=11.0.0=pypi_0
pip=24.2=py310h06a4308_0
platformdirs=4.3.6=pypi_0
pooch=1.8.2=pypi_0
portalocker=3.0.0=pypi_0
proglog=0.1.10=pypi_0
propcache=0.2.0=pypi_0
protobuf=5.28.3=pypi_0
psutil=6.1.0=pypi_0
pyarrow=17.0.0=pypi_0
pycparser=2.22=pypi_0
pynvml=11.5.3=pypi_0
pyparsing=3.2.0=pypi_0
pysocks=1.7.1=pypi_0
pysoundfile=0.9.0.post1=pypi_0
python=3.10.15=he870216_1
python-dateutil=2.9.0.post0=pypi_0
python-dotenv=1.0.1=pypi_0
pytorchvideo=0.1.5=pypi_0
pytz=2024.2=pypi_0
pyyaml=6.0.2=pypi_0
readline=8.2=h5eee18b_0
regex=2024.9.11=pypi_0
requests=2.32.3=pypi_0
safetensors=0.4.5=pypi_0
scikit-learn=1.6.0=pypi_0
scipy=1.14.1=pypi_0
sentencepiece=0.2.0=pypi_0
setuptools=75.1.0=py310h06a4308_0
six=1.16.0=pypi_0
soundfile=0.12.1=pypi_0
soupsieve=2.6=pypi_0
soxr=0.5.0.post1=pypi_0
sqlalchemy=2.0.37=pypi_0
sqlite=3.45.3=h5eee18b_0
sympy=1.13.1=pypi_0
tabulate=0.9.0=pypi_0
termcolor=2.5.0=pypi_0
threadpoolctl=3.5.0=pypi_0
timm=1.0.12=pypi_0
tk=8.6.14=h39e8969_0
tokenizers=0.19.1=pypi_0
torch=2.2.2+cu121=pypi_0
torchaudio=2.2.2+cu121=pypi_0
torchmetrics=1.6.1=pypi_0
torchvision=0.15.0=pypi_0
tqdm=4.66.5=pypi_0
transformers=4.41.2=pypi_0
triton=2.2.0=pypi_0
typeguard=2.13.3=pypi_0
typing-extensions=4.12.2=pypi_0
tzdata=2024.2=pypi_0
urllib3=2.2.3=pypi_0
wcwidth=0.2.13=pypi_0
wheel=0.44.0=py310h06a4308_0
xxhash=3.5.0=pypi_0
xz=5.4.6=h5eee18b_1
yacs=0.1.8=pypi_0
yarl=1.16.0=pypi_0
zipp=3.20.2=pypi_0
zlib=1.2.13=h5eee18b_1
